+++ dd-java-agent/instrumentation/play-2.3/play-2.3.gradle
@@ -8,7 +8,7 @@ muzzle {
   pass {
     group = 'com.typesafe.play'
     module = 'play_2.11'
-    versions = '[2.3.0,2.4)'
+    versions = '[2.3.9,2.4)'
     assertInverse = true
   }
   fail {
